requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
sqlite3
asyncio
aiohttp==3.9.1
cryptography==41.0.8